+++ prepForBattle.js
@@ -1,0 +1,110 @@
+export const $divArenas = document.querySelector('.arenas');
+
+export const $chat = document.querySelector('.chat');
+
+export const logs = {
+    start: 'Часы показывали [time], когда [player1] и [player2] бросили вызов друг другу.',
+    end: [
+        'Результат удара [playerWins]: [playerLose] - труп',
+        '[playerLose] погиб от удара бойца [playerWins]',
+        'Результат боя: [playerLose] - жертва, [playerWins] - убийца',
+    ],
+    hit: [
+        '[playerDefence] пытался сконцентрироваться, но [playerKick] разбежавшись раздробил копчиком левое ухо врага.',
+        '[playerDefence] расстроился, как вдруг, неожиданно [playerKick] случайно раздробил грудью грудину противника.',
+        '[playerDefence] зажмурился, а в это время [playerKick], прослезившись, раздробил кулаком пах оппонента.',
+        '[playerDefence] чесал <вырезано цензурой>, и внезапно неустрашимый [playerKick] отчаянно размозжил грудью левый бицепс оппонента.',
+        '[playerDefence] задумался, но внезапно [playerKick] случайно влепил грубый удар копчиком в пояс оппонента.',
+        '[playerDefence] ковырялся в зубах, но [playerKick] проснувшись влепил тяжелый удар пальцем в кадык врага.',
+        '[playerDefence] вспомнил что-то важное, но внезапно [playerKick] зевнув, размозжил открытой ладонью челюсть противника.',
+        '[playerDefence] осмотрелся, и в это время [playerKick] мимоходом раздробил стопой аппендикс соперника.',
+        '[playerDefence] кашлянул, но внезапно [playerKick] показав палец, размозжил пальцем грудь соперника.',
+        '[playerDefence] пытался что-то сказать, а жестокий [playerKick] проснувшись размозжил копчиком левую ногу противника.',
+        '[playerDefence] забылся, как внезапно безумный [playerKick] со скуки, влепил удар коленом в левый бок соперника.',
+        '[playerDefence] поперхнулся, а за это [playerKick] мимоходом раздробил коленом висок врага.',
+        '[playerDefence] расстроился, а в это время наглый [playerKick] пошатнувшись размозжил копчиком губы оппонента.',
+        '[playerDefence] осмотрелся, но внезапно [playerKick] робко размозжил коленом левый глаз противника.',
+        '[playerDefence] осмотрелся, а [playerKick] вломил дробящий удар плечом, пробив блок, куда обычно не бьют оппонента.',
+        '[playerDefence] ковырялся в зубах, как вдруг, неожиданно [playerKick] отчаянно размозжил плечом мышцы пресса оппонента.',
+        '[playerDefence] пришел в себя, и в это время [playerKick] провел разбивающий удар кистью руки, пробив блок, в голень противника.',
+        '[playerDefence] пошатнулся, а в это время [playerKick] хихикая влепил грубый удар открытой ладонью по бедрам врага.',
+    ],
+    defence: [
+        '[playerKick] потерял момент и храбрый [playerDefence] отпрыгнул от удара открытой ладонью в ключицу.',
+        '[playerKick] не контролировал ситуацию, и потому [playerDefence] поставил блок на удар пяткой в правую грудь.',
+        '[playerKick] потерял момент и [playerDefence] поставил блок на удар коленом по селезенке.',
+        '[playerKick] поскользнулся и задумчивый [playerDefence] поставил блок на тычок головой в бровь.',
+        '[playerKick] старался провести удар, но непобедимый [playerDefence] ушел в сторону от удара копчиком прямо в пятку.',
+        '[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.',
+        '[playerKick] не думал о бое, потому расстроенный [playerDefence] отпрыгнул от удара кулаком куда обычно не бьют.',
+        '[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.'
+    ],
+    draw: 'Ничья - это тоже победа!'
+};
+
+
+class Player {
+    constructor(props) {
+        this.player = props.player;
+        this.name = props.name;
+        this.hp = props.hp;
+        this.img = props.img;
+    }
+    attack = () => {
+        console.log(`${this.name} Fight...`)
+    };
+    
+    changeHP = (damage) => {
+        this.hp -= damage;
+    
+        if(this.hp <= 0) {
+            this.hp = 0;
+        }
+    };
+    elHP = () => {
+        return document.querySelector(`.player${this.player} .life`);
+    };
+    renderHP() {
+        const $playerLifeContainer = this.elHP(); 
+       $playerLifeContainer.style.width = this.hp + '%';
+    };
+}
+
+export const player1 = new Player({player: 1, name: 'Scorpion', hp: 100, img: 'http://reactmarathon-api.herokuapp.com/assets/scorpion.gif'});
+export const player2 = new Player({player: 2, name: 'Sub-zero', hp: 100, img: 'http://reactmarathon-api.herokuapp.com/assets/subzero.gif'})
+
+export function createElement(tag, className) {
+    const $tag = document.createElement(tag);
+
+   if(className) {
+    $tag.classList.add(className);
+   } 
+
+    return $tag;
+}
+
+
+export function createPlayer({player, name, hp, img }) {
+    const $playerDiv = createElement('div', `player${player}`);
+
+    const $playerProgressbar = createElement('div', 'progressbar');
+    $playerDiv.appendChild($playerProgressbar);
+
+    const $playerCharacter = createElement('div', 'character');
+    $playerDiv.appendChild($playerCharacter);
+
+    const $progressbarLife = createElement('div', 'life');
+    $progressbarLife.style.width = hp + '%';
+    $playerProgressbar.appendChild($progressbarLife);
+
+    const $progressbarName = createElement('div', 'name');
+    $progressbarName.innerText = name;
+    $playerProgressbar.appendChild($progressbarName);
+
+    const $characterImg = createElement('img');
+    $characterImg.src = img;
+    $playerCharacter.appendChild($characterImg);
+
+    return $playerDiv
+}
+
